requirements.txt
matplotlib==3.10.8
numpy==2.4.4
opencv_python==4.9.0.80
opencv_python_headless==4.13.0.92
Pillow==12.2.0
PyYAML==6.0.2
PyYAML==6.0.3
scikit_learn==1.8.0
scipy==1.17.1
segmentation_models_pytorch==0.5.0
torch==2.8.0
tqdm==4.67.1
wandb==0.26.0
